// src/shared/types/mock-server-data.ts
export type MockServerData = {
    titles: string[];
    descriptions: string[];
    postDates: string[];
    previewImages: string[];
    images: string[];
    housingTypes: string[];
    amenities: string[];
    userNames: string[];
    emails: string[];
    avatars: string[];
    userTypes: string[];
};